v0.2.2-debug
Fixed an issue from v0.2.1-debug

+++ src/net/client/filters.rs
@@ -61,8 +61,8 @@ pub enum FilterCode{
             FilterCode::GameDir => TypeId::of::<&&str>(),
             FilterCode::Map => TypeId::of::<&&str>(),
             FilterCode::GameAddr => TypeId::of::<&&str>(),
-            FilterCode::AppId => TypeId::of::<&u16>(),
-            FilterCode::NoAppId => TypeId::of::<&u16>(),
+            FilterCode::AppId => TypeId::of::<&u32>(),
+            FilterCode::NoAppId => TypeId::of::<&u32>(),
         }
     }
